requirements.txt
-i https://pypi.org/simple
amqp==5.1.1 ; python_version >= '3.6'
asgiref==3.6.0 ; python_version >= '3.7'
async-timeout==4.0.2 ; python_version >= '3.6'
billiard==3.6.4.0
boto3==1.26.46
botocore==1.29.46 ; python_version >= '3.7'
celery==5.2.7
certifi==2022.12.7 ; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==2.1.1 ; python_full_version >= '3.6.0'
click==8.1.3 ; python_version >= '3.7'
click-didyoumean==0.3.0 ; python_full_version >= '3.6.2' and python_full_version < '4.0.0'
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt[crypto]==5.2.2
drf-yasg==1.21.4
gunicorn==20.1.0
idna==3.4 ; python_version >= '3.5'
inflection==0.5.1 ; python_version >= '3.5'
itypes==1.2.0
jinja2==3.1.2 ; python_version >= '3.7'
jmespath==1.0.1 ; python_version >= '3.7'
kombu==5.2.4 ; python_version >= '3.7'
markupsafe==2.1.1 ; python_version >= '3.7'
packaging==23.0 ; python_version >= '3.7'
prompt-toolkit==3.0.36 ; python_full_version >= '3.6.2'
psycopg2-binary==2.9.5
pycparser==2.21
pyjwt==2.6.0 ; python_version >= '3.7'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-http-client==3.3.7 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2022.7
pyyaml==6.0
redis==4.4.1
requests==2.28.1
ruamel.yaml==0.17.21 ; python_version >= '3'
ruamel.yaml.clib==0.2.7 ; python_version < '3.11' and platform_python_implementation == 'CPython'
s3transfer==0.6.0 ; python_version >= '3.7'
sendgrid==6.9.7
sentry-sdk==1.12.1
setuptools==65.6.3 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.3 ; python_version >= '3.5'
starkbank-ecdsa==2.2.0
stripe==5.0.0
unidecode==1.3.6
uritemplate==4.1.1 ; python_version >= '3.6'
urllib3==1.26.13 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
vine==5.0.0 ; python_version >= '3.6'
watchdog==2.2.1
wcwidth==0.2.5
